add method for get and set brief of article

+++ src/main/java/repository/ArticleRepository.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 public interface ArticleRepository {
     Article create(Article article) throws SQLException;
 
-    void delete(int id) throws SQLException;
+    void delete(long id) throws SQLException;
 
   //  Article read(int id) throws SQLException;
 }
