+++ src/main/java/repository/UserRepository.java
@@ -7,8 +7,8 @@ import java.sql.SQLException;
 public interface UserRepository {
     //CRUD
     User create(User article) throws SQLException;
-    User read(int id) throws SQLException;
-    void delete(int id) throws SQLException;
+    User read(long id) throws SQLException;
+    void delete(long id) throws SQLException;
 
 }
